# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mublo AI Assistant API",
    "version": "0.2.0",
    "description": "Android AI Assistant, SaaS API, and private Worker contract. Business tenancy is always scoped by company_id, independently from Framework domain_id."
  },
  "servers": [
    { "url": "https://{host}/mublo-ai/api/v1", "variables": { "host": { "default": "crm.meccaphone.shop" } } }
  ],
...
  "paths": {
...
    "/messaging/suppressions/events": {
      "post": {
        "tags": ["Messaging policy"],
        "operationId": "appendSuppressionEvent",
        "description": "Appends a versioned SUPPRESS or LIFT event and updates the current suppression projection atomically. LIFT never restores contact permission.",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuppressionEvent" } } } },
        "responses": {
          "200": { "description": "Suppression event stored" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "opaque" },
      "workerToken": { "type": "apiKey", "in": "header", "name": "X-Worker-Token" }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
      },
...
    },
    "responses": {
...
      "Unauthorized": { "description": "Missing, expired, or revoked credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Forbidden": { "description": "Authenticated but outside the principal company or capability", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
      "Conflict": { "description": "Version, token rotation, or idempotency conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "ValidationError": { "description": "Semantic validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
    },
    "schemas": {
...
      "SuppressionEvent": { "$ref": "./schemas/suppression-event-v1.schema.json" },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["result", "success", "message", "error"],
        "properties": {
          "result": { "const": "error" },
          "success": { "const": false },
          "message": { "type": "string" },
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code"],
            "properties": {
              "code": { "type": "string", "pattern": "^[A-Z][A-Z0-9_]{2,63}$" },
              "details": { "type": "object" },
              "request_id": { "type": "string", "maxLength": 128 }
            }
          }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }]
}